+++ desktop/CodexWebRemote.Launcher/ServerManager.cs
@@ -58,6 +58,7 @@ internal sealed class ServerManager : IDisposable
             start.Environment["CODEX_WEB_SESSION_HOURS"] = "24";
             start.Environment["CODEX_WEB_DATA_DIR"] = _paths.DataRoot;
             start.Environment["CODEX_WEB_UPDATE_REQUEST_FILE"] = _paths.UpdateRequestFile;
+            start.Environment["CODEX_WEB_LAUNCHER_PATH"] = _paths.ExecutablePath;
             _process = new Process { StartInfo = start, EnableRaisingEvents = true };
             _process.OutputDataReceived += (_, e) => AppendLog(e.Data);
             _process.ErrorDataReceived += (_, e) => AppendLog(e.Data);